requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
fastf1==3.4.4
apache-airflow-providers-mongo==5.0.0